Update to include permissions required
It was confusing at first why the instructions were not accurate and my "Multi-Factor Authentication" button was inactive.  Buried in other documentation I believe this to be that I did not have proper role access.  Following any of the three actions to enable MFA on AD global administration is required.  This should be stated at the top of the document to save15 minutes of reading and trial/error attempting to make a change that is not possible.